Privatized Decomposition
