+++ projectq/cengines/_replacer/__init__.py
@@ -10,7 +10,6 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-from ._decomposition import Decomposition
 from ._decomposition_rule import DecompositionRule, ThisIsNotAGateClassError
 from ._decomposition_rule_set import DecompositionRuleSet
 from ._replacer import (AutoReplacer,
